[LayoutNG] Fixedpos static position in nested multicol w/ viewport CB
If a fixedpos was nested in an OOF inside a nested multicol, and the
OOF ancestor resulted in new columns being added to the inner
multicol, we would end up with the wrong static position if the
fixedpos' containing block was the viewport.

The reason being that we weren't writing the newly added columns back
to the related legacy flow thread, which is needed in order to
calculate the correct static position once we get to the viewport. To
fix this, extend the logical bottom of the final fragmentainer group
in NGOutOfFlowLayoutPart::LayoutOOFsInMulticol() if any new columns
are added.

Bug: 1200328
Change-Id: Ic9ddde46226ca7bb0d0a420427814cb46438de6d