Missing text node reattach for first-letter changes
Sometimes we have LayoutText which does not contain anything considered
to be a rendered first letter. For instance, if it only contains
punctuation. Modifying the character data to something which constitutes
a first letter will require a layout object reattach to correctly
(re-)construct the LayoutTextFragment for the rendered ::first-letter.

Bug: 1229931, 1214543
Change-Id: I5befff0d89b207321a1357076d823d1274dae6d8